Rework views with new Selects and xo item renderers
